third_party/blink/web_tests/http/tests/devtools/profiler/heap-profiler-profiling-expected.txt

Tests that sampling heap profiling works.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testProfiling
found pageFunction heap-profiler-profiling.js:16
found (anonymous) heap-profiler-profiling.js:17

Profiler was disabled.
